separated the components into another props comonent and called the props component in Our Goals
