+++ src/components/about/OurGoals.tsx
@@ -1,16 +1,8 @@
-import { StaticImageData } from "next/image";
-import Image from "next/image";
-
-type OurGoalsProps = {
-  image: StaticImageData;
-  text: string;
-};
-const OurGoals = ({ image, text }: OurGoalsProps) => {
+import OurGoalsProps from "@/components/about/OurGoalsProps";
+import myImage from "@/public/image.jpg"
+const OurGoals = () => {
   return (
-    <div className="mx-20 flex flex-col items-center justify-center border-y-4 border-ula-yellow-primary py-12 font-medium">
-      <Image src={image} alt="logo" />
-      {text}
-    </div>
+    <OurGoalsProps image={myImage} text="placeholder"/>
   );
 };
 export default OurGoals;
